Mejoré el estilo de las cards y las rediseñé para que sean reutilizables
