+++ src/components/Admin/Product/Table/TablaFilaProductos.jsx
@@ -4,7 +4,7 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { Link } from "react-router-dom";
 import { toast } from "sonner";
 import Swal from "sweetalert2";
-import { Button } from "primereact/button";
+import { Button } from 'primereact/button';
 
 import { deleteProductoFn } from "../../../../api/productos.js";
 import { useProducto } from "../../../../stores/useProducto.js";
@@ -12,13 +12,12 @@ import ModalProductos from "../ModalProductos.jsx";
 
 import "../styles/producto.css";
 
-const ProductCard = (props) => {
-  const { producto } = props;
+const TablaFilaProductos = (props) => {
+  const { producto, index } = props;
 
   const { setProductoToEdit } = useProducto();
 
   const [modalData, setModalData] = useState(null);
-  const [visible, setVisible] = useState(false);
 
   const queryClient = useQueryClient();
 
@@ -39,8 +38,8 @@ const ProductCard = (props) => {
   });
 
   const handleMoreInfo = () => {
-    setModalData(producto);
-    setVisible(true);
+    setModalData(producto); // Set the current product data
+    setVisible(true);       // Show the modal
   };
 
   const handleEdit = () => {
@@ -64,22 +63,32 @@ const ProductCard = (props) => {
   };
 
   return (
-    <div className="card single__product mb-4">
-      <div className="product__img">
-        <img
-          src={producto.imagen}
-          alt={producto.nombre}
-          className="h-100 w-100 img-fluid object-fit-cover"
-        />
-      </div>
-      <div className="product__content">
-        <h6 className="text-center fs-3 my-2">{producto.nombre}</h6>
-          <p className="text-center">{producto.descripcion}</p>
-        <hr />
-        <div className="d-flex justify-content-around align-items-center m-2">
-          <Button
-            className="btn btn-primary"
-            onClick={handleMoreInfo}
+    <>
+      <tr className="row-height">
+        <td className="h-100">{index + 1}</td>
+        <td>
+          <img
+            alt={producto.nombre}
+            className="productos-tabla-imagen img-fluid"
+            src={producto.imagen}
+          />
+        </td>
+        <td>{producto.nombre}</td>
+        <td>
+          {producto.habilitado === "true" ? (
+            <i className="bi bi-check-circle-fill text-success"></i>
+          ) : (
+            <i className="bi bi-x-circle-fill text-danger"></i>
+          )}
+        </td>
+        <td className="row-height text-end row row-size gap-2">
+          <Button icon="pi pi-external-link"
+            className="btn btn-primary col-auto"
+            to="#"
+            onClick={(e) => {
+              e.preventDefault();
+              handleMoreInfo();
+            }}
           >
             <i className="bi bi-plus-circle tamaño-icono"></i>
           </Button>
@@ -93,27 +102,23 @@ const ProductCard = (props) => {
           <button className="btn btn-danger col-auto" onClick={handleDelete}>
             <i className="bi bi-trash3-fill tamaño-icono"></i>
           </button>
-        </div>
-      </div>
+        </td>
+      </tr>
       {modalData && (
-        <ModalProductos
-          values={modalData}
-          visible={visible}
-          onHide={() => setVisible(false)}
-        />
+        <ModalProductos values={modalData} onClose={() => setModalData(null)} />
       )}
-    </div>
+    </>
   );
 };
 
-export default ProductCard;
+export default TablaFilaProductos;
 
-ProductCard.propTypes = {
+TablaFilaProductos.propTypes = {
   producto: PropTypes.shape({
     id: PropTypes.string.isRequired,
     nombre: PropTypes.string.isRequired,
     imagen: PropTypes.string.isRequired,
     descripcion: PropTypes.string.isRequired,
-    preciounitario: PropTypes.number.isRequired,
-  }).isRequired,
+  }),
+  index: PropTypes.number.isRequired,
 };
