+++ src/views/DetalleProductoView.jsx
@@ -1,4 +1,4 @@
-import FormularioProductos from "../components/Admin/Productos/FormularioProductos"
+import FormularioProductos from "../components/Admin/Product/FormularioProductos"
 
 const DetalleProductoView = () => {
   return (
