+++ src/views/routing/RootView.jsx
@@ -1,6 +1,6 @@
 import { Outlet, useLocation } from "react-router-dom";
 import Navbar from "../../components/Common/navBar/navbar";
-import '../../components/Admin/Productos/styles/producto.css';
+import '../../components/Admin/Product/styles/producto.css';
 
 const RootView = () => {
   const location = useLocation();
